requirements.txt
# python == 3.8.12
pyyaml == 5.4.1
matplotlib == 3.2.2
opencv-python == 4.5.4.60

numpy == 1.21.2
scipy == 1.7.3
torch == 1.9.0
onnx == 1.10.2
onnxruntime == 1.10.0

# qt
pyqt5 == 5.15.6
pyqt5-qt5 == 5.15.2
pyqt5-sip == 12.9.0

pycocotools == 2.0.3
